Add: モンスターのコンポーネントをTypeScript対応させた
