+++ frontend/src/components/Games/IntermediateMonster.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 // Colors
 import { COLORS } from '../../style_constants.js';
 
-import AdvancedMonsterImage from '../../images/advanced.png';
+import IntermediateMonsterImage from '../../images/intermediate.png';
 
 // handleColorType
 import { handleColorType } from '../../functions/handleColorType';
@@ -16,41 +16,28 @@ import {
   FadeOutAnime
 } from '../shared_style';
 
-const AdvancedWrapper = styled.div`
+import { MonsterArg } from '../../types/components/games';
+
+const IntermediateWrapper = styled.div<{gameResult: "" | "progress" | "win" | "lose"}>`
   animation-name: ${({ gameResult }) => gameResult === "win" && FadeOutAnime};
   animation-duration: 3s;
   animation-fill-mode: forwards;
   text-align: center;
-  position: relative;
 `;
 
 // 初登場時のモンスター
-const FirstAdvancedMonsterWrapper = styled.img`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translateY(-95%) translateX(-50%);
-  -webkit-transform: translateY(-95%) translateX(-50%);
-  margin: auto;
-  width: 50vw;
-  height: 55vh;
-  z-index: 2;
+const FirstIntermediateMonsterWrapper = styled.img`
+  width: 26%;
+  height: 26%;
   object-fit: contain;
   animation-name: ${FadeInAnime};
   animation-duration: 3s;
   animation-fill-mode: forwards;
 `;
 
-const AdvancedMonsterWrapper = styled.img`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translateY(-95%) translateX(-50%);
-  -webkit-transform: translateY(-95%) translateX(-50%);
-  margin: auto;
-  width: 50vw;
-  height: 55vh;
-  z-index: 2;
+const IntermediateMonsterWrapper = styled.img<{questionJudgement: "progress" | "correct" | "incorrect"}>`
+  width: 26%;
+  height: 26%;
   object-fit: contain;
   animation-name: ${({ questionJudgement }) => questionJudgement === "correct" && MonsterFlash};
   animation-duration: 1s;
@@ -58,8 +45,8 @@ const AdvancedMonsterWrapper = styled.img`
   animation-iteration-count: 1;
 `;
 
-const HpGageWrapper = styled.div`
-  width: 50vw;
+const HpGageWrapper = styled.div<{firstAppearance: boolean}>`
+  width: 40%;
   border-radius: 3px;
   background-color: ${COLORS.LIGHT_BLACK};
   animation-name: ${({ firstAppearance }) => firstAppearance && FadeInAnime};
@@ -70,7 +57,7 @@ const HpGageWrapper = styled.div`
   margin: 0 auto;
 `;
 
-const InnerHpGageWrapper = styled.div`
+const InnerHpGageWrapper = styled.div<{monsterMaxHp: number, monsterHp: number}>`
   width: ${({ monsterHp, monsterMaxHp }) => `${100 * (monsterHp / monsterMaxHp)}%`};
   transition: 0.5s;
   height: 1.8vh;
@@ -81,29 +68,29 @@ const InnerHpGageWrapper = styled.div`
   box-shadow: 0 2px 2px 0 rgba(255,255,255,.2) inset,0 2px 10px 0 rgba(255,255,255,.5) inset,0 -2px 2px 0 rgba(0,0,0,.1) inset;
 `;
 
-export const AdvancedMonster = ({
+export const IntermediateMonster = ({
   monsterHp,
   monsterMaxHp,
   questionJudgement,
   firstAppearance,
   gameResult,
   gameDescriptionOpen
-}) => {
+}: MonsterArg): JSX.Element => {
   return (
     <>
       {
         !gameDescriptionOpen &&
-          <AdvancedWrapper
+          <IntermediateWrapper
             gameResult={gameResult}
           >
             {
               firstAppearance ?
-                <FirstAdvancedMonsterWrapper
-                  src={AdvancedMonsterImage}
+                <FirstIntermediateMonsterWrapper
+                  src={IntermediateMonsterImage}
                 />
               :
-                <AdvancedMonsterWrapper
-                  src={AdvancedMonsterImage}
+                <IntermediateMonsterWrapper
+                  src={IntermediateMonsterImage}
                   questionJudgement={questionJudgement}
                 />
             }
@@ -115,7 +102,7 @@ export const AdvancedMonster = ({
                 monsterMaxHp={monsterMaxHp}
               />
             </HpGageWrapper>
-          </AdvancedWrapper>
+          </IntermediateWrapper>
 
       }
     </>
